requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pyngrok==7.0.5
